remap.apply_func
Apply a function like np.mean or np.max to neighborhood around all
points. It is like inverse distance weighting with power=0 but there is
no weighting. All points in neighborhood are treated the same.

The neighborhood is at a fixed radius instead of a fixed number of
neighbors.

Eliminated depreciated destination_obj argument.