Add spatial hashing (#1169)
I attempted to stay close to the KD-Tree and BallTree API. However, the
purpose of the SpatialHash class is to help locate the faces that a
list of coordinates lie within. Because of this, the intent of the
`query` function for the SpatialHash is a bit different than that for
the KD and Ball trees.

Additional support routines are provided for assessing whether a point
is inside or outside a polygon. This is done by calculating the
barycentric coordinates of a point in a convex polygon. The method is
defined so that the sum of the barycentric coordinates is exactly one.
Because of this, if any of the barycentric coordinates are negative, we
immediately know that a point is outside the polygon.

All coordinate calculations are done using (lon,lat) in radians for the
SpatialHash class. Cartesian coordinates are not supported in this
commit.

* Add `get_spatial_hash` to grid api

* Add example usage in spatial hashing docs

* Add spatialhashing to userguide

* Add basic tests

* Fix call to get_spatialhash

* Fix formatting

* Prepend attributes intended for internal use with _

* Add a few patches for hashtable setup and query check

* Ensure that the lon bounds are sorted from low to high. This is needed
to make sure that the index looping for filling the hashtable actually
executes.

* All barycentric coordinates are calculated directly. Rather than
calculating the last coordinate to ensure they sum to one, we now
compute all coordinates. The coordinates are calculated now to align
with the indexing of the node ID's so that np.sum(bcoords*nodes) returns
the barycentric interpolation estimate of point. If the point is within
the element, then the barycentric interpolation estimate matches the
input coordinate (to machine precision). I've added this as a check.

* Fix docstrings for tests; add fesom test with particle list

* Add docstring for get_spatialhash

* remove empty notes

* Update uxarray/grid/grid.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Fix naming in test

* Update uxarray/grid/neighbors.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Update uxarray/grid/neighbors.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Update test/test_spatial_hashing.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Fix docstring for reconstruct field

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Update uxarray/grid/grid.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Add example to docstring

* Use waschpress points for generalized barycentric coordinates

Added dual and primal grids for mpas. The generalized waschpress points
give us the correct generalized barycentric coordinates that satisfy the
Lagrange property for barycentric interpolation in convex polygons.

* Switch to mpas example in notebook

* Update uxarray/grid/grid.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Retain output in notebook. Make requested formatting changes

* Fix formatting

* Change to smaller/simpler 4xhex grid

* Add get_spatial_hash to the Grid/Methods section

* Clear outputs from notebook.

* Fix docstring; we're not computing signed area any longer.

Instead, we're favoring a check of the predicted coordinates using
barycentric interpolation

* Update uxarray/grid/neighbors.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Update uxarray/grid/neighbors.py

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>

* Add query optimization suggested by @philipc2

* Remove extra return statement

* Clip denominator in barycentric weights calculation

The denominator in the barycentric weights calculation (Weischell
weights) is clipped to be no smaller than `ERROR_TOLERANCE` from the
uxarray.constants module.

This change is made to fix an issue, reported anectdotally by @philip2c
in https://github.com/UXARRAY/uxarray/pull/1169

* Clip triangle areas rather than denominator

When the denominator is clipped, rather than the individual triangle
areas, the Weischell weights are incorrect for the case when a query
point is on a face vertex.

I've added tests for query points on edges and vertices

---------

Co-authored-by: Philip Chmielowiec <67855069+philipc2@users.noreply.github.com>